backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
firebase-admin
pydantic
